requirements.txt


streamlit>=1.28
google-generativeai